Move the play trailer button to a separate file
